#16 Migrate test to tests folder
