--- src/tests/mock/yaml_structure.py
@@ -1,3 +1,0 @@
-from pandas import DataFrame 
-
-yaml_structure = [{'dataframe': [DataFrame({'tudo_junto': ['124234 EIRUWOEJI 345-345', '124234 EIRUWOEJI 345-345', '124234 EIRUWOEJI 345-345']}), DataFrame({'apenas_number': ['124234', '124234', '124234']})], 'data_lang': 'es', 'dimensions': {'tudo_junto': 23, 'apenas_number': 25}, 'enabled_features': {'ksmash_sequence_vowels': True, 'ksmash_sequence_consonants': True, 'ksmash_sequence_special_characters': True, 'ksmash_numbers': True, 'ksmash_char_frequence': True, 'word_embedding': True}}, {'dataframe': [DataFrame({'apenas_number':['124234', '124234', '124234']})], 'data_lang': 'es', 'dimensions': None, 'enabled_features': {'ksmash_sequence_vowels': False, 'ksmash_sequence_consonants': True, 'ksmash_sequence_special_characters': True, 'ksmash_numbers': True, 'ksmash_char_frequence': True, 'word_embedding': False}}]
