+++ tests/mock/success/yaml_mock.yaml
@@ -2,7 +2,7 @@ dag:
   id: basic_example
   description: DAG de teste
   data_path:
-    - 'src/tests/mock/data_mock.csv'
+    - 'tests/mock/data_mock.csv'
   output_folder: 'output'
 
   feature_engineering:
